+++ bin/test.sh
@@ -3,12 +3,6 @@
 # exit when any command fails
 set -e
 
-mvn clean -Pspark-3.2 -Pscala-2.12
-mvn test -Pspark-3.2 -Pscala-2.12
-
-mvn clean -Pspark-3.2 -Pscala-2.13
-mvn test -Pspark-3.2 -Pscala-2.13
-
 mvn clean -Pspark-3.3 -Pscala-2.12
 mvn test -Pspark-3.3 -Pscala-2.12
 
@@ -20,3 +14,9 @@ mvn test -Pspark-3.4 -Pscala-2.12
 
 mvn clean -Pspark-3.4 -Pscala-2.13
 mvn test -Pspark-3.4 -Pscala-2.13
+
+mvn clean -Pspark-3.5 -Pscala-2.12
+mvn test -Pspark-3.5 -Pscala-2.12
+
+mvn clean -Pspark-3.5 -Pscala-2.13
+mvn test -Pspark-3.5 -Pscala-2.13